### 2022-10-11 #4841 @editorialbot
[PRE REVIEW]: RingsPy: A Python package for Voronoi mesh generation of cellular solids with radial growth pattern
**Submitting author:** @kingyin3613 (<a href="http://orcid.org/0000-0003-4921-9786">Hao Yin</a>)
**Repository:** https://github.com/kingyin3613/RingsPy
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.5
