### 2022-10-11 990f9d6
Update DOIs for JOSS paper and README
