+++ contents/paper.bib
@@ -6,6 +6,7 @@
   number={17},
   pages={R878--R882},
   year={2017},
+  DOI={10.1017/CBO9781139878326},
   publisher={Elsevier}
 }
 
@@ -17,6 +18,7 @@
   number={7},
   pages={1658},
   year={2020},
+  DOI={10.3390/cells9071658},
   publisher={MDPI}
 }
 
@@ -28,6 +30,7 @@
   number={3},
   pages={261--272},
   year={2020},
+  DOI={10.1017/CBO9781139878326},
   publisher={Nature Publishing Group}
 }
 @article{saye2011voronoi,
@@ -38,6 +41,7 @@
   number={49},
   pages={19498--19503},
   year={2011},
+  DOI={10.1073/pnas.1111557108},
   publisher={National Acad Sciences}
 }
 
@@ -48,6 +52,7 @@
   volume={35},
   number={1},
   pages={77--88},
+  DOI={10.15252/embj.201592374},
   year={2016}
 }
 
@@ -59,10 +64,12 @@
   number={4},
   pages={1--12},
   year={2016},
+  DOI={10.1145/2897824.2925922},
   publisher={ACM New York, NY, USA}
 }
 
 @book{gibson_ashby_1997, place={Cambridge}, edition={2}, series={Cambridge Solid State Science Series}, title={Cellular Solids: Structure and Properties}, DOI={10.1017/CBO9781139878326}, publisher={Cambridge University Press}, author={Gibson, Lorna J. and Ashby, Michael F.}, year={1997}, collection={Cambridge Solid State Science Series}}
+
 @article{harris2020array,
   title={Array programming with NumPy},
   author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
@@ -71,6 +78,7 @@
   number={7825},
   pages={357--362},
   year={2020},
+  DOI={10.1017/CBO9781139878326},
   publisher={Nature Publishing Group}
 }
 @article{liu2022growth,
@@ -81,6 +89,7 @@
   number={6609},
   pages={975--981},
   year={2022},
+  DOI={10.1126/science.abn1459},
   publisher={American Association for the Advancement of Science}
 }
 
@@ -92,6 +101,7 @@
   number={1},
   pages={1--15},
   year={2017},
+  DOI={10.1007/s13595-016-0613-y},
   publisher={Springer}
 }
 
@@ -103,6 +113,7 @@
   number={4},
   pages={1813--1863},
   year={2019},
+  DOI={10.1007/s00332-019-09531-w},
   publisher={Springer}
 }
 
@@ -114,6 +125,7 @@
   number={2},
   pages={109--138},
   year={2011},
+  DOI={10.1016/j.pmatsci.2010.06.001},
   publisher={Elsevier}
 }
 
@@ -125,5 +137,6 @@
   number={1},
   pages={19},
   year={2019},
+  DOI={10.3390/designs3010019},
   publisher={MDPI}
 }
