+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = RingsPy
-version = 0.2.6
+version = 0.2.7
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
 description = A geometric generation tool for prismatic cellular solids
