### 2022-10-11 #4841 @kingyin3613
> @kingyin3613 thanks for this submission. Can you please check those missing DOIs? ☝️ You can call `@editorialbot check references` to check them again.

@Kevin-Mattheus-Moerman thanks for the prompt comment! I have checked these DOIs, and maybe now it is bot's turn
