+++ tests/test_hexalattice_hex.py
@@ -23,15 +23,17 @@ from pathlib import Path
 def test_hexalattice_hex():
     # performance check only
     startTime = time.time()
-
+    
     # ==================================================================
     # Input parameters
     geoName = 'honeycomb_hex'
     path = 'meshes'
-
+    
+    radial_growth_rule = 'regular_hexagonal'
     iter_max = 100
     print_interval = 500
-
+    
+    # Radial cell growth parameters
     # length unit: mm
     r_min = 0   # inner radius of wood log
     r_max = 2   # outer radius of wood log
@@ -40,59 +42,84 @@ def test_hexalattice_hex():
     width_early = 0.7*(r_max-r_min)/nrings # early wood ring width
     width_late = 0.3*(r_max-r_min)/nrings # late wood ring width
     log_center = (0,0) # coordinates of log center in the global system of reference
-    box_center = (0,0) # coordinates of box center in the global system of reference
-    box_size   = 1 # side length of the clipping box
     
     cellsize_early = 0.2
     cellsize_late = 0.2
     cellwallthickness_early = 0.010
-    cellwallthickness_late = 0.006
-    cellangle = 0.0
-
+    cellwallthickness_late = 0.010
+    
+    # clipping box parameters
+    boundaryFlag = 'on'
+    box_shape = 'hexagon'
+    box_center = (0,0) # coordinates of box center in the global system of reference
+    box_size = 1.0 # side length
+        
+    # longitudinal direction parameters
+    fiberlength = 0.5*box_size
+    theta_min = 0 # radian
+    theta_max = 0 # radian
+    z_min = 0
+    z_max = box_size
+    long_connector_ratio = 0.02 # longitudinal joint size
+    
     merge_operation = 'off'
     merge_tol = 0.01
-
+    
     precrackFlag = 'off'
     precrack_widths = 0.1
-
-    boundaryFlag = 'on'
-    STLFlag = 'on'
-
+    
+    stlFlag = 'on'
+    
+    inpFlag = 'on'
+    inpType = 'Abaqus'
+    
+    # ==================================================================
+    # Remove directory/files if exists already
+    try:
+        shutil.rmtree(Path(path, geoName))
+    except:
+        pass
+    
+    if not os.path.exists(Path(path, geoName)):
+        os.makedirs(Path(path, geoName))
+        
+    # ==================================================================
+    # Random cell site placement with a radial growth pattern
+    
+    if radial_growth_rule == 'wood_binary':
+        # ---------------------------------------------
+        # wood with binary annual ring pattern (earlywood-latewood, abrupt transition)
+        sites,radius = RPgen.CellPlacement_Wood(log_center,r_max,r_min,nrings,width_heart,\
+                            width_early,width_late,cellsize_early,cellsize_late,\
+                            iter_max,print_interval)
+    elif radial_growth_rule == 'regular_hexagonal':
+        # ----------------------------------
+        # hexagonal honeycomb lattice
+        sites,radius = RPgen.CellPlacement_Honeycomb(log_center,r_max,r_min,nrings,\
+                            box_center,box_size,width_heart,\
+                            width_early,width_late,\
+                            cellsize_early,cellsize_late,\
+                            cellwallthickness_early,cellwallthickness_late,\
+                            iter_max,print_interval)
+    elif os.path.splitext(radial_growth_rule)[1] == '.npy':
+        # ----------------------------------
+        # load saved cell sites and radius data
[…]
