+++ tests/test_wood_cube.py
@@ -27,12 +27,14 @@ def test_wood_cube():
 
     # ==================================================================
     # Input parameters
-    geoName = 'wood_cube'
+    geoName = 'test_cube'
     path = 'meshes'
 
+    radial_growth_rule = 'wood_binary'   
     iter_max = 100
     print_interval = 500
-
+    
+    # Radial cell growth parameters
     # length unit: mm
     r_min = 0   # inner radius of wood log
     r_max = 4   # outer radius of wood log
@@ -41,14 +43,85 @@ def test_wood_cube():
     width_early = 0.7*(r_max-r_min)/nrings # early wood ring width
     width_late = 0.3*(r_max-r_min)/nrings # late wood ring width
     log_center = (0,0) # coordinates of log center in the global system of reference
+
+    cellsize_early = 0.02
+    cellsize_late = 0.01
+    cellwallthickness_early = 0.010
+    cellwallthickness_late = 0.006
+    
+    # clipping box parameters
+    boundaryFlag = 'on'
+    box_shape = 'square'
     box_center = (1.25,0) # coordinates of box center in the global system of reference
-    box_size = 1.0 # cube size
-
-    x_min = box_center[0] - box_size/2
-    x_max = box_center[0] + box_size/2 
-    y_min = box_center[1] - box_size/2
-    y_max = box_center[1] + box_size/2 
-
+    box_size = 1.0 # side length
+        
+    # longitudinal direction parameters
+    fiberlength = 0.5*box_size
+    theta_min = 0 # radian
+    theta_max = 0.05 # radian
+    z_min = 0
+    z_max = box_size
+    long_connector_ratio = 0.02 # longitudinal joint size
+    
+    merge_operation = 'on'
+    merge_tol = 0.01
+    
+    precrackFlag = 'off'
+    precrack_widths = 0.1
+    
+    stlFlag = 'on'
+    
+    inpFlag = 'on'
+    inpType = 'Abaqus'
+    
+    # ==================================================================
+    # Remove directory/files if exists already
+    try:
+        shutil.rmtree(Path(path, geoName))
+    except:
+        pass
+    
+    if not os.path.exists(Path(path, geoName)):
+        os.makedirs(Path(path, geoName))
+        
+    # ==================================================================
+    # Random cell site placement with a radial growth pattern
+    
+    if radial_growth_rule == 'wood_binary':
+        # ---------------------------------------------
+        # wood with binary annual ring pattern (earlywood-latewood, abrupt transition)
+        sites,radius = RPgen.CellPlacement_Wood(log_center,r_max,r_min,nrings,width_heart,\
+                            width_early,width_late,cellsize_early,cellsize_late,\
+                            iter_max,print_interval)
+    elif radial_growth_rule == 'regular_hexagonal':
+        # ----------------------------------
+        # hexagonal honeycomb lattice
+        sites,radius = RPgen.CellPlacement_Honeycomb(log_center,r_max,r_min,nrings,\
+                            box_center,box_size,width_heart,\
+                            width_early,width_late,\
+                            cellsize_early,cellsize_late,\
+                            cellwallthickness_early,cellwallthickness_late,\
+                            iter_max,print_interval)
+    elif os.path.splitext(radial_growth_rule)[1] == '.npy':
+        # ----------------------------------
+        # load saved cell sites and radius data
+        sites,radius = RPgen.ReadSavedSites(radial_growth_rule)
+    
+    else:
[…]
