+++ tests/test_read_existing_sites.py
@@ -19,8 +19,10 @@ import time
 import RingsPy.MeshGenTools as RPgen
 import shutil
 from pathlib import Path
+import pytest
 
-
+@pytest.mark.datafiles('/test_radius.npy')
+@pytest.mark.datafiles('/test_sites.npy')
 def test_read_existing_sites():
     # performance check only
     startTime = time.time()
