+++ RingsPy.egg-info/requires.txt
@@ -1,11 +1,0 @@
-requests>=2
-numpy
-scipy
-matplotlib>=3.2.2
-hexalattice
-vtk
-
-[testing]
-pytest>=6.0
-pytest-cov>=2.0
-tox>=3.24
