+++ RingsPy.egg-info/SOURCES.txt
@@ -1,14 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-setup.cfg
-setup.py
-RingsPy/MeshGenTools.py
-RingsPy/__init__.py
-RingsPy/py.typed
-RingsPy.egg-info/PKG-INFO
-RingsPy.egg-info/SOURCES.txt
-RingsPy.egg-info/dependency_links.txt
-RingsPy.egg-info/not-zip-safe
-RingsPy.egg-info/requires.txt
-RingsPy.egg-info/top_level.txt
