+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = RingsPy
-version = 0.3.0
+version = 0.3.2
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
 description = A geometric generation tool for prismatic cellular solids
@@ -55,6 +55,7 @@ zip_safe = no
 testing =
     pytest>=6.0
     pytest-cov>=2.0
+	pytest-datafiles>=2.0
     tox>=3.24
 
 [options.package_data]
