### 2022-10-24 02cd2c8
Delete test_read_existing_sites.py
