+++ tests/test_read_existing_sites.py
@@ -1,297 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-RingsPy
-a Python Voronoi diagrams-based geometric generation tool that generates 
-3D meshes of prismatic cellular solids with radial growth rules.
-
-author: king_yin3613
-email: haoyin2022@u.northwestern.edu
-"""
-
-import numpy as np
-from scipy.spatial import Voronoi, voronoi_plot_2d
-from scipy.spatial import Delaunay, delaunay_plot_2d
-import matplotlib.pyplot as plt
-import matplotlib.patches as patches
-import matplotlib.tri as mtri
-import os
-import time
-import RingsPy.MeshGenTools as RPgen
-import shutil
-from pathlib import Path
-import pytest
-
-@pytest.mark.datafiles('/test_radius.npy')
-@pytest.mark.datafiles('/test_sites.npy')
-def test_read_existing_sites():
-    # performance check only
-    startTime = time.time()
-    
-    # ==================================================================
-    # Input parameters
-    geoName = 'wood_cube'
-    path = 'meshes'
-    
-    radial_growth_rule = 'test_sites.npy'   
-    iter_max = 100
-    print_interval = 500
-    
-    # Radial cell growth parameters
-    # length unit: mm
-    r_min = 0   # inner radius of wood log
-    r_max = 4   # outer radius of wood log
-    nrings = 4 # number of rings
-    width_heart = 0.3*(r_max-r_min)/nrings # heart wood ring width
-    width_early = 0.7*(r_max-r_min)/nrings # early wood ring width
-    width_late = 0.3*(r_max-r_min)/nrings # late wood ring width
-    log_center = (0,0) # coordinates of log center in the global system of reference
-    
-    cellsize_early = 0.02
-    cellsize_late = 0.01
-    cellwallthickness_early = 0.010
-    cellwallthickness_late = 0.006
-    
-    # clipping box parameters
-    boundaryFlag = 'on'
-    box_shape = 'square'
-    box_center = (1.25,0) # coordinates of box center in the global system of reference
-    box_size = 1.5 # side length
-        
-    # longitudinal direction parameters
-    fiberlength = 0.5*box_size
-    theta_min = 0 # radian
-    theta_max = 0 # radian
-    z_min = 0
-    z_max = box_size
-    long_connector_ratio = 0.02 # longitudinal joint size
-    
-    merge_operation = 'on'
-    merge_tol = 0.01
-    
-    precrackFlag = 'off'
-    precrack_widths = 0.1
-    
-    stlFlag = 'on'
-    
-    inpFlag = 'on'
-    inpType = 'Abaqus'
-    
-    # ==================================================================
-    # Remove directory/files if exists already
-    try:
-        shutil.rmtree(Path(path, geoName))
-    except:
-        pass
-    
-    if not os.path.exists(Path(path, geoName)):
-        os.makedirs(Path(path, geoName))
-        
-    # ==================================================================
-    # Random cell site placement with a radial growth pattern
-    
-    if radial_growth_rule == 'wood_binary':
-        # ---------------------------------------------
-        # wood with binary annual ring pattern (earlywood-latewood, abrupt transition)
-        sites,radius = RPgen.CellPlacement_Wood(log_center,r_max,r_min,nrings,width_heart,\
-                            width_early,width_late,cellsize_early,cellsize_late,\
-                            iter_max,print_interval)
-    elif radial_growth_rule == 'regular_hexagonal':
-        # ----------------------------------
-        # hexagonal honeycomb lattice
-        sites,radius = RPgen.CellPlacement_Honeycomb(log_center,r_max,r_min,nrings,\
[…]
