### 2022-11-16 54840d3
preparing v0.3.3
Add conda-forge setup option, add a test for reading existing mesh file
