+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = RingsPy
-version = 0.3.2
+version = 0.3.3
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
 description = A geometric generation tool for prismatic cellular solids
@@ -36,14 +36,14 @@ install_requires =
     requests>=2
 	numpy
 	scipy
-	matplotlib >= 3.2.2
+	matplotlib-base >= 3.2.2
 	hexalattice
 	vtk
 setup_requires =
     requests>=2
 	numpy
 	scipy
-	matplotlib >= 3.2.2
+	matplotlib-base >= 3.2.2
 	hexalattice
 	vtk
 python_requires = >=3.6
