+++ setup.cfg
@@ -36,14 +36,14 @@ install_requires =
     requests>=2
 	numpy
 	scipy
-	matplotlib-base >= 3.2.2
+	matplotlib>=3.2.2
 	hexalattice
 	vtk
 setup_requires =
     requests>=2
 	numpy
 	scipy
-	matplotlib-base >= 3.2.2
+	matplotlib>=3.2.2
 	hexalattice
 	vtk
 python_requires = >=3.6
