### 2022-11-17 ea43702
Try put external data files in RingsPy directory
