### 2022-11-18 #4841 @Kevin-Mattheus-Moerman
( @mahtab-vafaee is a PhD student in the group of Dr Ted Vaughan at the University of Galway, who do research on trabecular bone lattice architectures. )
