### 2022-11-18 #4945 @kingyin3613
Dear all, thanks for agreeing to review my paper! May I request, to review version v0.3.4 (latest version) of RingsPy? In that version, I attached a few new functionalities including saving/reading generated Voronoi cell nuclei, etc. I appreciate that in advance! (@Kevin-Mattheus-Moerman what else should I do in order to update the software version in the reviewing process?)
