### 2022-12-21 kingyin3613/ringspy#1 @niallmadden
Does not appear to install correctly.
I tried to install this on a fresh conda env:

> $ conda create --name ringspy pip
> $ pip install RingsPy [OK]
> <clone https://github.com/kingyin3613/RingsPy>
> $ pytest .
>

Errors:
```
rootdir: /home/<personal>/RingsPy, configfile: pyproject.toml
plugins: hydra-core-1.2.0, cov-3.0.0, mock-3.8.2, anyio-3.6.1
collected 0 items / 4 errors
/usr/local/lib/python3.8/dist-packages/coverage/control.py:794: CoverageWarning: No data was collected. (no-data-collected)
  self._warn("No data was collected.", slug="no-data-collected")

================================================= ERRORS ==================================================
_____________________________ ERROR collecting tests/test_hexalattice_hex.py ______________________________
ImportError while importing test module '/home/<personal>/RingsPy/tests/test_hexalattice_hex.py'.
Traceback:
/usr/lib/python3.8/importlib/__init__.py:127: in import_module
tests/test_hexalattice_hex.py:19: in <module>
    import RingsPy.MeshGenTools as RPgen
E   ModuleNotFoundError: No module named 'RingsPy'

```
