### 2022-12-21 #4945 @niallmadden
This does not work for me: "pip install" without error, but tests fail - No module named 'RingsPy'
https://github.com/kingyin3613/RingsPy/issues/1#issue-1505490200
