### 2022-12-21 kingyin3613/ringspy#1 @kingyin3613
@niallmadden Thanks for letting me know this issue.

I think this happened because the package was installed in the wrong environment. It is like `pip install` installs the packages within Python environment and `conda install` installs the packages within conda environment, and you are using conda environment indeed. See this thread for more details: https://stackoverflow.com/questions/56134588/is-that-a-bad-idea-to-use-conda-and-pip-install-on-the-same-environment.

However, I did publish RingsPy to conda-forge channel, so a better way to install is, use the following command to install this package directly from conda:
> conda install -c conda-forge ringspy

and then
> pytest .
