### 2022-12-22 kingyin3613/ringspy#1 @niallmadden
I'm familar with that issue; installing pip within conda usually resolves it.
I created a new conda environment, and installed with "conda install -c conda-forge ringspy"
I observed the same issue as above. Minimally:

$ "python -c "import rinspy"
> Traceback (most recent call last):
>   File "<string>", line 1, in <module>
> ModuleNotFoundError: No module named 'rinspy

Same result when I deactivate conda, and just use pip install; installs without error, but still can't be found.
Even tried in docker container.
