### 2022-12-23 kingyin3613/ringspy#1 @kingyin3613
I tried create a new conda environment and install within that environment as well and it seems that use of capitalized package name worked fine. Try this:

> "python -c "import RingsPy"

This is weird, but please let me know if this works for you. I will try to fix this problem for the newer releases.
