### 2023-01-06 kingyin3613/ringspy#1 @niallmadden
OK; that worked. Thanks! Would be good to change to more standard module-naming convention: https://peps.python.org/pep-0008/#package-and-module-names

FYI, "tests$ pytest ." fails for test_read_existing_sites.py; passes the other 3.
