### 2023-02-03 #4945 @niallmadden
> @Kevin-Mattheus-Moerman I had a problem with the installation. It's now working and I'm running/checking the codes.

Same here, and informed the authors. [https://github.com/kingyin3613/RingsPy/issues/1](https://github.com/kingyin3613/RingsPy/issues/1)
Working now, though the standard convention of using only lower-case for module names is not followed.
