+++ contents/paper.md
@@ -31,7 +31,7 @@ Natural and bioinspired materials are often characterized by irregular and heter
 
 Cellular yet irregular morphology of materials is a natural outcome of local cell growth that follows simple intrinsic physical rules, among these, the Voronoi diagram, has been proposed to be one of the best approximations of cell interfaces at the steady-state by the equilibrium of intracellular pressures and the intercellular surface tension [@saye2011voronoi;@sanchez2016fundamental]. Besides, there is also an underlying process, which governs the global growth of cells to form the radial pattern in plane and the growth of either straight or helical grains in the perpendicular direction to the plane [@hartmann2017modelling;@sozio2019nonlinear]. Understanding the role of locally irregular and globally radial morphology in determining material properties offers a new path to unveil the behavior of natural materials and engineer materials with superior functionalities, such as anisotropic programmable material properties, stress redirection, and oriented impact energy dissipation. 
 
-While the laboratory and in-site imaging approaches (e.g., 3D X-ray tomography) can be reliable for reproducing the most geometrically accurate material morphology, however, these approaches are still costly and time-consuming in some highly repetitive scenarios, such as prototyping and probabilistic analyses. Given that, an algorithm-based, stochastic generator of cellular geometries is at a new level of need. 
+While the laboratory and in-situ imaging approaches (e.g., 3D X-ray tomography) can be reliable for reproducing the most geometrically accurate material morphology, however, these approaches are still costly and time-consuming in some highly repetitive scenarios, such as prototyping and probabilistic analyses. Given that, an algorithm-based, stochastic generator of cellular geometries is at a new level of need. 
 
 We propose a fundamental, growth-inspired program that evokes the formation of cellular architectures with radial and longitudinal growth patterns in natural systems. This virtual growth program imposes both local and global generation rules on a limited category of basic elements. It generates meshes that exhibit stochastic geometrical properties, starting from very limited initial constraints, which echoes the diversity of biological and biomimetic systems. 
 
