+++ tests/test_hexalattice_hex.py
@@ -16,7 +16,7 @@ import matplotlib.patches as patches
 import matplotlib.tri as mtri
 import os
 import time
-import RingsPy.MeshGenTools as RPgen
+import ringspy.MeshGenTools as rpgen
 import shutil
 from pathlib import Path
 
@@ -89,13 +89,13 @@ def test_hexalattice_hex():
     if radial_growth_rule == 'wood_binary':
         # ---------------------------------------------
         # wood with binary annual ring pattern (earlywood-latewood, abrupt transition)
-        sites,radius = RPgen.CellPlacement_Wood(log_center,r_max,r_min,nrings,width_heart,\
+        sites,radius = rpgen.CellPlacement_Wood(log_center,r_max,r_min,nrings,width_heart,\
                             width_early,width_late,cellsize_early,cellsize_late,\
                             iter_max,print_interval)
     elif radial_growth_rule == 'regular_hexagonal':
         # ----------------------------------
         # hexagonal honeycomb lattice
-        sites,radius = RPgen.CellPlacement_Honeycomb(log_center,r_max,r_min,nrings,\
+        sites,radius = rpgen.CellPlacement_Honeycomb(log_center,r_max,r_min,nrings,\
                             box_center,box_size,width_heart,\
                             width_early,width_late,\
                             cellsize_early,cellsize_late,\
@@ -104,7 +104,7 @@ def test_hexalattice_hex():
     elif os.path.splitext(radial_growth_rule)[1] == '.npy':
         # ----------------------------------
         # load saved cell sites and radius data
-        sites,radius = RPgen.ReadSavedSites(radial_growth_rule)
+        sites,radius = rpgen.ReadSavedSites(radial_growth_rule)
     
     else:
         print('Growth rule: {:s} is not supported for the current version, please check the README for more details.'.format(radial_growth_rule))
@@ -119,7 +119,7 @@ def test_hexalattice_hex():
     # ==================================================================
     # Clipping box (Boundaries) of the model
     x_min,x_max,y_min,y_max,boundaries,boundary_points,boundarylines = \
-        RPgen.Clipping_Box(box_shape,box_center,box_size,boundaryFlag)
+        rpgen.Clipping_Box(box_shape,box_center,box_size,boundaryFlag)
     
     # if precracked
     x_indent_size = box_size*0.120
@@ -154,7 +154,7 @@ def test_hexalattice_hex():
         [voronoi_vertices,finite_ridges,boundary_points,finite_ridges_new,\
          boundary_ridges_new,nvertex,nvertices_in,nfinite_ridge,nboundary_ridge,\
          nboundary_pts,nboundary_pts_featured,voronoi_ridges,nridge] = \
-            RPgen.RebuildVoronoi_merge(vor,sites,boundaries,log_center,x_min,x_max,y_min,y_max,box_center,box_shape,merge_tol,boundaryFlag)
+            rpgen.RebuildVoronoi_merge(vor,sites,boundaries,log_center,x_min,x_max,y_min,y_max,box_center,box_shape,merge_tol,boundaryFlag)
         
         RebuildvorTime = time.time() 
         print('Voronoi tessellation rebuilt and merged in {:.3f} seconds'.format(RebuildvorTime - voronoiTime))
@@ -162,7 +162,7 @@ def test_hexalattice_hex():
         [voronoi_vertices,finite_ridges,boundary_points,finite_ridges_new,\
           boundary_ridges_new,nvertex,nvertices_in,nfinite_ridge,nboundary_ridge,\
           nboundary_pts,nboundary_pts_featured,voronoi_ridges,nridge] = \
[…]
