### 2023-02-04 b2273f6
Update setup.cfg
