+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = RingsPy
+name = ringspy
 version = 0.3.5
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
@@ -8,7 +8,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 license = GPL-3.0
 license_files = LICENSE
-url = https://github.com/kingyin3613/RingsPy/
+url = https://github.com/kingyin3613/ringspy/
 platforms = unix, linux, osx, cygwin, win32
 python_requires = ">=3.6"
 
