+++ RingsPy/__init__.py
@@ -4,5 +4,5 @@
 import sys
 
 
-__version__ = '0.2.0'
+__version__ = '0.3.5'
 VERSION = __version__
