+++ RingsPy/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import sys
-
-
-__version__ = '0.3.5'
-VERSION = __version__
