### 2023-02-05 ef2ae20
Delete and push ringspy folder to rename it with all small letters
