### 2023-02-05 615ddc5
change ci settings as Python 3.6 is not supported on Ubuntu 22.04
