### 2023-02-05 #4945 @mahtab-vafaee
> > @Kevin-Mattheus-Moerman I had a problem with the installation. It's now working and I'm running/checking the codes.
>
> Hi Mahtab@mahtab-vafaee , could you please inform me what the issue was with the installation? I'll try fix it with the issue mentioned by @niallmadden asap. Thanks!

The problem was with Conda installation on my system, resolved now. Thanks.
