### 2023-02-16 #4945 @niallmadden
I've given this package some time. It meets most of the criteria. And the latest version installs OK via conda, and follows the standard naming convention. But it fails at the tests. For reprodicibiltiy, here is what I tried:
$ conda create --name ringspy
$ conda activate ringspy
$ conda install -c conda-forge ringspy
$ git clone git@github.com:kingyin3613/ringspy.git
$ cd ringspy/tests/
$  pytest .
Output includes:
est_hexalattice_hex.py .                                                                                                            [ 25%]
test_hexalattice_tri.py .                                                                                                            [ 50%]
test_read_existing_sites.py F                                                                                                        [ 75%]
test_wood_cube.py .                                                                                                                  [100%]
.
.
Could not find file: test_sites.npy, please check if the existing site file is under the same directory with the input script.
Now exitting...
---------- coverage: platform linux, python 3.8.10-final-0 -----------
Name                                                                     Stmts   Miss  Cover
--------------------------------------------------------------------------------------------
/home/niall/.local/lib/python3.8/site-packages/ringspy/MeshGenTools.py    2110    623    70%
/home/niall/.local/lib/python3.8/site-packages/ringspy/__init__.py           3      0   100%
--------------------------------------------------------------------------------------------
TOTAL                                                                     2113    623    71%

========================================================= short test summary info ==========================================================
FAILED test_read_existing_sites.py::test_read_existing_sites - SystemExit: None
======================================================= 1 failed, 3 passed in 5.02s ========================================================

---

So, either this package is just not quite right, or it needs a reviewer who can better get around installation issues.
