+++ tests/test_hexalattice_hex.py
@@ -30,38 +30,42 @@ def test_hexalattice_hex():
     path = 'meshes'
     
     radial_growth_rule = 'regular_hexagonal'
-    iter_max = 100
-    print_interval = 500
+    iter_max = 500 # increase this number to achieve a more regular geometry
+    print_interval = 500 # interval for printing prgress info
     
     # Radial cell growth parameters
     # length unit: mm
-    r_min = 0   # inner radius of wood log
-    r_max = 2   # outer radius of wood log
+    r_min = 0   # inner radius of generation domain
+    r_max = 2   # outer radius of generation domain
     nrings = 4 # number of rings
-    width_heart = 0.3*(r_max-r_min)/nrings # heart wood ring width
-    width_early = 0.7*(r_max-r_min)/nrings # early wood ring width
-    width_late = 0.3*(r_max-r_min)/nrings # late wood ring width
-    log_center = (0,0) # coordinates of log center in the global system of reference
+    width_heart = 0.3*(r_max-r_min)/nrings # ring width for the innermost ring
+    width_sparse = 0.7*(r_max-r_min)/nrings # ring width for rings with sparse cells
+    width_dense = 0.3*(r_max-r_min)/nrings # ring width for rings with dense cells
+    generation_center = (0,0) # coordinates of generation domain center
     
-    cellsize_early = 0.2
-    cellsize_late = 0.2
-    cellwallthickness_early = 0.010
-    cellwallthickness_late = 0.010
+    cellsize_sparse = 0.2
+    cellsize_dense = 0.2
+    cellwallthickness_sparse = 0.010
+    cellwallthickness_dense = 0.010
     
     # clipping box parameters
     boundaryFlag = 'on'
     box_shape = 'hexagon'
-    box_center = (0,0) # coordinates of box center in the global system of reference
+    box_center = (0,0) # coordinates of clipping box center
     box_size = 1.0 # side length
         
     # longitudinal direction parameters
-    fiberlength = 0.5*box_size
-    theta_min = 0 # radian
-    theta_max = 0 # radian
+    segment_length = 0.5*box_size
+    theta_min = 0 # unit: radian
+    theta_max = 0 # unit: radian
     z_min = 0
     z_max = box_size
-    long_connector_ratio = 0.02 # longitudinal joint size
+    long_connector_ratio = 0.02 # longitudinal joint length = ratio * segment_length
     
+    # material parameters
+    skeleton_density = 1.5e-9 # unit: tonne/mm3
+    
+    # generation parameters
     merge_operation = 'off'
     merge_tol = 0.01
     
@@ -84,27 +88,35 @@ def test_hexalattice_hex():
         os.makedirs(Path(path, geoName))
         
     # ==================================================================
-    # Random cell site placement with a radial growth pattern
+    # Place cells with a specific radial growth pattern
     
-    if radial_growth_rule == 'wood_binary':
+    if radial_growth_rule == 'binary':
         # ---------------------------------------------
-        # wood with binary annual ring pattern (earlywood-latewood, abrupt transition)
-        sites,radius = rpgen.CellPlacement_Wood(log_center,r_max,r_min,nrings,width_heart,\
-                            width_early,width_late,cellsize_early,cellsize_late,\
+        # binary radial growth pattern (e.g. wood microstructure with earlywood-latewood alternations)
+        sites,radii = rpgen.CellPlacement_Binary(generation_center,r_max,r_min,nrings,width_heart,\
+                            width_sparse,width_dense,cellsize_sparse,cellsize_dense,\
                             iter_max,print_interval)
+    elif radial_growth_rule == 'binary_lloyd':
+        # ---------------------------------------------
[…]
