+++ ringspy/__init__.py
@@ -4,5 +4,5 @@
 import sys
 
 
-__version__ = '0.3.5'
+__version__ = '0.4.0'
 VERSION = __version__
