### 2023-02-20 510b2a4
fix the test for reading existing sites and minor update for README
