+++ ringspy/__init__.py
@@ -4,5 +4,5 @@
 import sys
 
 
-__version__ = '0.4.0'
+__version__ = '0.4.1'
 VERSION = __version__
