### 2023-02-20 #4945 @kingyin3613
> ## I've given this package some time. It meets most of the criteria. And the latest version installs OK via conda, and follows the standard naming convention. But it fails at the tests. For reprodicibiltiy, here is what I tried:
> $ conda create --name ringspy
> $ conda activate ringspy
> $ conda install -c conda-forge ringspy
> $ git clone [git@github.com](mailto:git@github.com):kingyin3613/ringspy.git
> $ cd ringspy/tests/
> $ pytest .
> Output includes:
> est_hexalattice_hex.py . [ 25%]
> test_hexalattice_tri.py . [ 50%]
> test_read_existing_sites.py F [ 75%]
> test_wood_cube.py . [100%]
> .
> .
> Could not find file: test_sites.npy, please check if the existing site file is under the same directory with the input script.
> Now exitting...
> ---------- coverage: platform linux, python 3.8.10-final-0 -----------
> Name Stmts Miss Cover
> ## /home/niall/.local/lib/python3.8/site-packages/ringspy/MeshGenTools.py 2110 623 70%
> /home/niall/.local/lib/python3.8/site-packages/ringspy/**init**.py 3 0 100%
> TOTAL 2113 623 71%
>
> ========================================================= short test summary info ========================================================== FAILED test_read_existing_sites.py::test_read_existing_sites - SystemExit: None ======================================================= 1 failed, 3 passed in 5.02s ========================================================
>
> So, either this package is just not quite right, or it needs a reviewer who can better get around installation issues.

 Sorry for the late reply @niallmadden . Actually, in the past few weeks, I have been working on resolving the conflicts between pytest and my way of storing external data files, which caused the issue you mentioned in this thread. I finally figured out an easy way to fix the issue, and the changes were published with the new features in the latest release (release v0.4.1). I have pytested on many different environments for the latest release, it turned out that all 4 old tests plus 1 new test were passed successfully. You may pull the latest version and retry the testing, please let me know if you find anything strange or unclear. I highly appreciate your patience and time in reviewing this software!
