### 2023-02-20 #4945 @niallmadden
>  Sorry for the late reply @niallmadden . Actually, in the past few weeks, I have been working on resolving the conflicts between pytest and my way of storing external data files, which caused the issue you mentioned in this thread. I finally figured out an easy way to fix the issue, and the changes were published with the new features in the latest release (release v0.4.1). I have pytested on many different environments for the latest release, it turned out that all 4 old tests plus 1 new test were passed successfully. You may pull the latest version and retry the testing, please let me know if you find anything strange or unclear. I highly appreciate your patience and time in reviewing this software!

@kingyin3613 Ok - I tested it again, using conda as above, and also with virtualenv+pip. All tests passed.
