### 2023-02-20 7dead20
add timeout to the while loop in ReadSavedSites in MeshGenTools.py
