### 2023-02-20 #4945 @kingyin3613
> Hi, I created an issue somewhere that I shouldn't and the issue is now closed by JOSS. Anyway, you can still see my comments [there](https://github.com/openjournals/joss-reviews/issues/5163).
>
> You might already have considered the issues in your latest update.

Hi @mahtab-vafaee , I have read that issue, and found it helpful for my future improvement!

The while loop in the "ReadSavedSites" function https://github.com/kingyin3613/ringspy/blob/615ddc5d07dd239c3383deac99755d23fdd76427/ringspy/MeshGenTools.py#L3921 , I also realized that there was a forever loop in reading external data; my thought was, as long as users store the external data in the same directory as the input scripts, they should be good. Adding a timeout condition here is a brilliant idea and I will do it.

The if condition https://github.com/kingyin3613/ringspy/blob/615ddc5d07dd239c3383deac99755d23fdd76427/ringspy/MeshGenTools.py#L3945 was a typo, it should also check if the radii file exists.

I have fixed both problems and committed a new version, you may try the [commit 7dead20](https://github.com/kingyin3613/ringspy/commit/7dead2028b6f63ed8ea5208ae50e5cdf62733dc4) now, let me know if you have any questions. Many thanks!
