### 2023-02-23 #4945 @niallmadden
> @niallmadden thanks for your help with this review. Since all boxes are ticked, are you happy to recommend acceptance in JOSS?

@Kevin-Mattheus-Moerman Yes, I confirm that I recommend acceptance.
