+++ tests/test_lloyd.py
@@ -55,11 +55,11 @@ def test_lloyd():
     box_size = 1.5 # side length
         
     # longitudinal direction parameters
-    segment_length = 0.5*box_size
+    nsegments = 2 # increase this number to achieve a smoother transition in longitudinal direction if theta is large
     theta_min = 0 # unit: radian
     theta_max = 0.05 # unit: radian
     z_min = 0
-    z_max = box_size
+    z_max = box_size # segment_length = (z_max - z_min) / nsegments
     long_connector_ratio = 0.02 # longitudinal joint length = ratio * segment_length
     
     # material parameters
@@ -197,16 +197,14 @@ def test_lloyd():
     ###############################################################################
     # Extrude in the parallel-to-grain (longitudinal) direction
     NURBS_degree = 2
-    nctrlpt_per_beam = 5
-    
     [IGAvertices,vertex_connectivity,beam_connectivity_original,nbeam_total,\
-     beam_connectivity,nbeamElem,nlayer,connector_t_connectivity,\
+     beam_connectivity,nbeamElem,nctrlpt_per_beam,nlayers,segment_length,connector_t_connectivity,\
      connector_t_bot_connectivity,connector_t_top_connectivity,\
      connector_t_reg_connectivity,connector_l_connectivity,nconnector_t_per_beam,\
      nconnector_t_per_grain,nconnector_t,nconnector_l,nconnector_total,\
-     theta,z_coord,nbeam_per_grain,connector_l_vertex_dict] = \
-        rpgen.GenerateBeamElement(NURBS_degree,nctrlpt_per_beam,\
-                            segment_length,theta_min,theta_max,z_min,z_max,\
+     theta,z_coord,connector_l_vertex_dict] = \
+        rpgen.GenerateBeamElement(NURBS_degree,nsegments,\
+                            theta_min,theta_max,z_min,z_max,\
                             long_connector_ratio,npt_per_layer,voronoi_vertices,\
                             nvertex,voronoi_ridges,nridge,generation_center,\
                             all_vertices_2D,max_wings,flattened_all_vertices_2D,all_ridges)
@@ -221,7 +219,7 @@ def test_lloyd():
         [precrack_elem,nconnector_t_precrack,nconnector_l_precrack] = \
             rpgen.insert_precracks(all_pts_2D,all_ridges,nridge,npt_per_layer,\
                                      npt_per_layer_normal,npt_per_layer_vtk,\
-                                     nlayer,precrack_nodes,precrack_widths,\
+                                     nlayers,precrack_nodes,precrack_widths,\
                                      cellsize_sparse)
     else:
         precrack_nodes = []
@@ -231,13 +229,12 @@ def test_lloyd():
     
     # ==================================================================
     # Calculate mesh info
-    height_connector_t = segment_length/4
-    
-    ConnMeshData = rpgen.ConnectorMeshFile(geoName,IGAvertices,connector_t_bot_connectivity,\
+    [ConnMeshData,conn_l_tangents] = \
+        rpgen.ConnectorMeshFile(geoName,IGAvertices,connector_t_bot_connectivity,\
                      connector_t_reg_connectivity,connector_t_top_connectivity,\
-                     height_connector_t,connector_l_connectivity,all_vertices_2D,\
-                     max_wings,flattened_all_vertices_2D,nbeam_per_grain,nridge,\
-                     connector_l_vertex_dict)
+                     connector_l_connectivity,all_vertices_2D,\
+                     max_wings,flattened_all_vertices_2D,nsegments,segment_length,\
+                     nctrlpt_per_beam,theta,nridge,connector_l_vertex_dict)
     
[…]
