### 2023-02-27 #4945 @kingyin3613
> > > Hi, I created an issue somewhere that I shouldn't and the issue is now closed by JOSS. Anyway, you can still see my comments [there](https://github.com/openjournals/joss-reviews/issues/5163).
> > > You might already have considered the issues in your latest update.
> >
> >
> > Hi @mahtab-vafaee , I have read that issue, and found it helpful for my future improvement!
> > The while loop in the "ReadSavedSites" function https://github.com/kingyin3613/ringspy/blob/615ddc5d07dd239c3383deac99755d23fdd76427/ringspy/MeshGenTools.py#L3921 , I also realized that there was a forever loop in reading external data; my thought was, as long as users store the external data in the same directory as the input scripts, they should be good. Adding a timeout condition here is a brilliant idea and I will do it.
> > The if condition https://github.com/kingyin3613/ringspy/blob/615ddc5d07dd239c3383deac99755d23fdd76427/ringspy/MeshGenTools.py#L3945 was a typo, it should also check if the radii file exists.
> > I have fixed both problems and committed a new version, you may try the [commit 7dead20](https://github.com/kingyin3613/ringspy/commit/7dead2028b6f63ed8ea5208ae50e5cdf62733dc4) now, let me know if you have any questions. Many thanks!
>
> Thanks for the update. All the tests pass. Just two technical issues regarding the longitudinal extrusion of the cells: 1- with a higher `theta_max` values, it could be better to use higher number of layers of cell extrusions to make this angle change smooth in transition to the next adjacent layer. For example, in the images below, with a higher `theta_max` on Fig. (a), you can see there's a big step between two neighboring layers, since I could create only limited number of longitudinal layers. Whereas with a lower angle gradient with respect to the z axis, like in Fig. (b), this transition is smoother, without big outliner corners. ![image](https://user-images.githubusercontent.com/64662505/221211732-53d28d17-0efd-48b3-9f18-0a26f7f10b34.png) For a `fiberlength` smaller than `0.4*box_size`, in the [test_hexalattice_hex.py](https://github.com/kingyin3613/ringspy/blob/main/tests/test_hexalattice_hex.py), for a `theta_max = 1.2`, an index error for the matrix [Meshdata](https://github.com/kingyin3613/ringspy/blob/7dead2028b6f63ed8ea5208ae50e5cdf62733dc4/ringspy/MeshGenTools.py#L1579) appears.
>
> 2- Also, since the `fiberlength`, is a factor of `box_size`, layers could overlap (Fig (c)), or there could be a large gap between them (Fig (d)). If this is considered as one of the features of the meshing tool, then it's ok. But, in the context of a solid mesh, to avoid this issue, I'd recommend to define the `fiberlength`, as a function of a new parameter of "Number of longitudinal layers", to divide the whole length into equal length elongated fiber sections.
>
> ![image](https://user-images.githubusercontent.com/64662505/221220569-f92a4daf-f341-4abe-b9be-7a760a9ae461.png)
>
> Hope this would help to improve your meshing tool.

Hi @mahtab-vafaee , thanks for your feedback and suggestions! I agree with you that using "number of longitudinal layers" instead of "fiber length" would be helpful to avoid the generation of non-physical solids. So, in the latest version, I changed the logic for the longitudinal extrusion accordingly, now users define "number of segments" instead of "segment length". You may pull the latest version (v0.4.2) or see [commit 5300740](https://github.com/kingyin3613/ringspy/commit/5300740216a5d4d0c4c4a214f4befad9b6079078). See below the effects of new mesh generation: a) `nsegments = 8` and `theta_max = 1.2`, b) `nsegments = 24` and `theta_max = 1.2`
![comparison](https://user-images.githubusercontent.com/17798082/221645357-0b6d6963-7b72-46ba-b321-4b0705fd5366.png)

Feel free to let me know if you find any other issues or have any questions!
