### 2023-03-06 b284dde
preparing for acceptance in JOSS
