+++ README.md
@@ -30,13 +30,13 @@ conda install -c conda-forge ringspy
 
 ### 3. Installation Check
 
-There are some unit tests in [tests](https://github.com/kingyin3613/ringspy/tree/main/tests/). One can use ``pytest`` to check whether the installation is successful by running this command:
+There are some unit tests in [tests](https://github.com/kingyin3613/ringspy/tree/main/tests/). One can use ``pytest`` to check whether the installation is successful by firstly moving to ``pytest`` directory and then running this command:
 
 ```bash
 pytest .
 ```
 
-## Getting Started (Current version: v0.4.2)
+## Getting Started (Current version: v1.0.0)
 Once all required components are installed and one is ready to begin, a path forward should be established for generating the mesh. The basic steps for running/viewing a cellular mesh are listed as the following:
 
     1. Edit geometry and algorithm parameters
@@ -51,7 +51,7 @@ The first step to generate a cellular geometry is selecting geometry and appropr
 ### 1.1. Geometry
 A template file, for example, `test_wood_cube.py` located in the [tests](https://github.com/kingyin3613/ringspy/tree/main/tests/) directory acts as both the parameter input file, and main executable for the generation of a cubic wood specimen.
 
-*Note: The Mesh Generation Tool by now only accepts many of pre-defined boundary geometries (for v0.4.x, the following 3 shapes are supported: triangle, square, hexagon), importing of CAD and/or other 3D model files will be implemented in subsequent versions.*
+*Note: The Mesh Generation Tool by now only accepts many of pre-defined boundary geometries (for v1.0.x, the following 3 shapes are supported: triangle, square, hexagon), importing of CAD and/or other 3D model files will be implemented in subsequent versions.*
 
 *Note: for greatest compatibility create the geometry using all millimeters.*
 
@@ -113,7 +113,7 @@ inpType = 'Abaqus'
 ```
 
 - `geoName` is the geometry name, `path` is the folder where the mesh files will be generated.
-- `radial_growth_rule` is the radial growth rule for cell placement. When a file name with extension`.npy` is specified, a saved cell data file will be loaded (for v0.4.x, choose one of these rules: `binary`, `binary_lloyd`, `regular_hexagonal`, or a file name with extension `.npy`).
+- `radial_growth_rule` is the radial growth rule for cell placement. When a file name with extension`.npy` is specified, a saved cell data file will be loaded (for v1.0.x, choose one of these rules: `binary`, `binary_lloyd`, `regular_hexagonal`, or a file name with extension `.npy`).
 - `iter_max` is the max number of iteration for randomly placing non-overlapping cell particles in the 2D toroidal cell placement region. Noticing that, larger `iter_max` leads to more centroidal Voronoi cells, for more reference, see [Wiki - Centroidal Voronoi Tessellation](https://en.wikipedia.org/wiki/Centroidal_Voronoi_tessellation/).
 - `print_interval` is the print interval when every n cell particles are placed in the placement region.
 - `r_min` and `r_max` are the upper and lower bounds of radii of toroidal cell placement regions (generation rings), `nrings` is the total number of rings.
@@ -121,7 +121,7 @@ inpType = 'Abaqus'
 - `generation_center` is the location of the placement region.
 - `cellsize_sparse`,`cellsize_dense`, `cellwallthickness_sparse`, and `cellwallthickness_dense` are parameters for the sparse and dense cells.
 - `boundaryFlag` flag can be turned on/off for generating neat boundaries consisting of grains.
-- `box_shape` is the shape of cutting box (for v0.4.x, choose one of following shapes: `triangle`, `square`, or `hexagon`).
[…]
