+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ringspy
-version = 0.4.2
+version = 1.0.0
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
 description = A geometric generation tool for prismatic cellular solids
