+++ ringspy/__init__.py
@@ -4,5 +4,5 @@
 import sys
 
 
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 VERSION = __version__
