+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ringspy
-version = 1.0.0
+version = 1.0.1
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
 description = A geometric generation tool for prismatic cellular solids
