### 2023-03-07 d2eb2de
update the package build configurations
