+++ .github/workflows/tests.yml
@@ -1,4 +1,4 @@
-name: Tests on Win-MacOS-Ubuntu with Python 3.6-3.10
+name: Tests on Win-MacOS-Ubuntu with Python 3.7-3.11
 
 on: [push, pull_request]
 
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-20.04, windows-latest, macos-latest]
-        python-version: ['3.6', '3.7', '3.8', '3.9', '3.10']
+        python-version: ['3.7', '3.8', '3.9', '3.10', '3.11']
 
     steps:
     - uses: actions/checkout@v2
