+++ tox.ini
@@ -1,16 +1,15 @@
 [tox]
 minversion = 3.8.0
-envlist = py36, py37, py38, py39, py310
+envlist = py37, py38, py39, py310, py311
 isolated_build = true
 
 [gh-actions]
 python =
-    3.6: py36
     3.7: py37
     3.8: py38
     3.9: py39
     3.10: py310
-
+    3.11: py311
 [testenv]
 setenv =
     PYTHONPATH = {toxinidir}
