+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ringspy
-version = 1.0.1
+version = 1.0.2
 author = Hao Yin
 author_email = haoyin2022@u.northwestern.edu
 description = A geometric generation tool for prismatic cellular solids
@@ -10,16 +10,16 @@ license = GPL-3.0
 license_files = LICENSE
 url = https://github.com/kingyin3613/ringspy/
 platforms = unix, linux, osx, cygwin, win32
-python_requires = ">=3.6"
+python_requires = ">=3.7"
 
 classifiers =
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
 	Programming Language :: Python :: 3.10
+	Programming Language :: Python :: 3.11
 	Development Status :: 4 - Beta
 	Intended Audience :: Science/Research
 	License :: OSI Approved :: GNU General Public License v3 (GPLv3)
@@ -46,7 +46,7 @@ setup_requires =
 	matplotlib>=3.2.2
 	hexalattice
 	vtk
-python_requires = >=3.6
+python_requires = >=3.7
 package_dir =
     
 zip_safe = no
