+++ .github/workflows/publish.yml
@@ -23,10 +23,11 @@ jobs:
     - name: Install dependencies
       run: |
         pip install setuptools wheel twine
+        pip install -q build
     - name: Publish
       env:
         TWINE_USERNAME: __token__
         TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
       run: |
-        python setup.py sdist bdist_wheel
+        python -m build
         twine upload dist/*
