### 2023-03-07 0c17c9b
Temporarily restrict the version of hexalattice to 1.2.1
