+++ setup.cfg
@@ -37,14 +37,14 @@ install_requires =
 	numpy
 	scipy
 	matplotlib>=3.2.2
-	hexalattice
+	hexalattice=1.2.1
 	vtk
 setup_requires =
     requests>=2
 	numpy
 	scipy
 	matplotlib>=3.2.2
-	hexalattice
+	hexalattice=1.2.1
 	vtk
 python_requires = >=3.7
 package_dir =
